+++ src/lib.rs
@@ -20,6 +20,8 @@ mod alloc_impl;
 #[cfg(feature = "arrayvec")]
 mod arrayvec_impl;
 pub mod cache;
+#[cfg(feature = "heapless-09")]
+mod heapless_09_impl;
 #[cfg(feature = "heapless")]
 mod heapless_impl;
 mod item;